+++ src/main/java/shopping/ui/AuthController.java
@@ -2,30 +2,12 @@ package shopping.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
-import shopping.application.AuthService;
-import shopping.dto.LoginRequest;
-import shopping.dto.LoginResponse;
 
 @Controller
 public class AuthController {
 
-    private final AuthService authService;
-
-    public AuthController(AuthService authService) {
-        this.authService = authService;
-    }
-
     @GetMapping("/login")
     public String login() {
         return "login";
-    }
-
-    @ResponseBody
-    @PostMapping("/login")
-    public LoginResponse login(@RequestBody LoginRequest loginRequest) {
-        return authService.login(loginRequest);
     }
 }
