+++ src/main/java/shopping/ui/AuthRestController.java
@@ -1,0 +1,23 @@
+package shopping.ui;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import shopping.application.AuthService;
+import shopping.dto.LoginRequest;
+import shopping.dto.LoginResponse;
+
+@RestController
+public class AuthRestController {
+
+    private final AuthService authService;
+
+    public AuthRestController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login")
+    public LoginResponse login(@RequestBody LoginRequest loginRequest) {
+        return authService.login(loginRequest);
+    }
+}
